+++ lib/ngtcp2_conn.h
@@ -289,7 +289,7 @@ struct ngtcp2_conn {
  *     Same packet number has already been added.
  */
 int ngtcp2_conn_sched_ack(ngtcp2_conn *conn, uint64_t pkt_num, int active_ack,
-                          ngtcp2_tstamp ts, uint8_t unprotected);
+                          ngtcp2_tstamp ts, int unprotected);
 
 /*
  * ngtcp2_conn_find_stream returns a stream whose stream ID is
